requirements.txt
# Core dependencies
python-dotenv==1.0.0
gradio==3.45.0

# Audio processing
pydub==0.25.1
speechrecognition==3.10.0
pyaudio==0.2.13

# Groq API
groq==0.4.0

# ElevenLabs API
elevenlabs==0.2.0

# Text-to-speech
gTTS==2.3.2

# Other utilities
platformdirs==3.11.0